Remove file IO stuff (moved to colugomusic/blahdio)

+++ include/snd/types.h
@@ -5,7 +5,7 @@
 namespace snd {
 
 using ChannelCount = std::uint16_t;
-using FrameCount = std::uint32_t;
+using FrameCount = std::uint64_t;
 using SampleRate = std::uint32_t;
 using BitDepth = std::uint16_t;
